\name{atoms.convertbasis}
\alias{atoms.convertbasis}
\title{Calculates basis converted atoms}
\usage{
  atoms.convertbasis(atoms, basis, forth = T)
}
\arguments{
  \item{atoms}{dataframe of atoms}

  \item{basis}{to use}

  \item{forth}{from cartesian to direct}
}
\description{
  \code{atoms.convertbasis} calculates basis converted
  atoms.
}

