\name{atoms.sort}
\alias{atoms.sort}
\title{Sorts atoms}
\usage{
  atoms.sort(atoms, sortindices = c(7, 3, 1, 2))
}
\arguments{
  \item{atoms}{dataframe of atoms}

  \item{sortindices}{vector of indices after which to sort.
  \code{c(1,3)} will sort by rx then by rz}
}
\description{
  \code{atoms.sort} sorts atoms.
}

