\name{bandsdata.addsympoint}
\alias{bandsdata.addsympoint}
\title{Adds custom high symmetry points to bandsdata}
\usage{
  bandsdata.addsympoint(bandsdata, indices)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{indices}{of high symmetry points in
  \code{bandsdata$kpoints}}
}
\description{
  \code{bandsdata.addsympoint} adds custom high symmetry
  points to bandsdata. returns class of type bandsdata
}

