\name{bandsdata.fit.dirac.function}
\alias{bandsdata.fit.dirac.function}
\title{Fit function for dirac fit}
\usage{
  bandsdata.fit.dirac.function(parameters, kpoints,
    constants)
}
\arguments{
  \item{parameters}{list containing all paramters}

  \item{kpoints}{indices of kpoints to use}

  \item{constants}{for use in fitfunction}
}
\description{
  \code{bandsdata.fit.dirac.function} contains fit function
  for dirac fit.
}

