\name{bandsdata.fit.dirac.makeconstants}
\alias{bandsdata.fit.dirac.makeconstants}
\title{Creates constants for dirac fit}
\usage{
  bandsdata.fit.dirac.makeconstants(bandsdata, bands,
    k0 = 0, factor = 1)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{bands}{indices of bands forming upper and lower
  dirac cone}

  \item{k0}{index of high symmetry point where the dirac
  cone appears}

  \item{factor}{\code{-1}/\code{1} for lower/upper dirac
  cone}
}
\description{
  \code{bandsdata.fit.dirac.makeconstants} creates
  constants for dirac fit.
}

