\name{bandsdata.fit.dirac.makeparameters}
\alias{bandsdata.fit.dirac.makeparameters}
\title{Creates startingparamters for dirac fit}
\usage{
  bandsdata.fit.dirac.makeparameters(vF)
}
\arguments{
  \item{vF}{starting Fermi velocity}
}
\description{
  \code{bandsdata.fit.dirac.makeparameters} creates
  startingparamters for dirac fit.
}

