\name{bandsdata.fit.quadratic.function}
\alias{bandsdata.fit.quadratic.function}
\title{Fit function for quadratic fit}
\usage{
  bandsdata.fit.quadratic.function(parameters, kpoints,
    constants)
}
\arguments{
  \item{parameters}{list containing all paramters}

  \item{kpoints}{indices of kpoints to use}

  \item{constants}{for use in fitfunction}
}
\description{
  \code{bandsdata.fit.quadratic.function} contains fit
  function for quadratic fit. used in
  \code{\link{bandsdata.fit}}.
}

