\name{bandsdata.fit.quadratic.makeconstants}
\alias{bandsdata.fit.quadratic.makeconstants}
\title{Creates constants for quadratic fit}
\usage{
  bandsdata.fit.quadratic.makeconstants(bandsdata, band,
    k0 = 0)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{band}{index of band to use}

  \item{k0}{index of high symmetry point to use}
}
\description{
  \code{bandsdata.fit.quadratic.makeconstants} creates
  constants for quadratic fit.
}

