\name{bandsdata.fit.quadratic.makeparameters}
\alias{bandsdata.fit.quadratic.makeparameters}
\title{Creates startingparamters for quadratic fit}
\usage{
  bandsdata.fit.quadratic.makeparameters(m)
}
\arguments{
  \item{m}{starting mass}
}
\description{
  \code{bandsdata.fit.quadratic.makeparameters} creates
  startingparamters for quadratic fit. used in
  \code{\link{bandsdata.fit}}.
}

