\name{bandsdata.getbanddistance}
\alias{bandsdata.getbanddistance}
\title{Gives the distance between to bands at one kpoint}
\usage{
  bandsdata.getbanddistance(bandsdata, kpoint, bands)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{kpoint}{index of kpoint to use}

  \item{bands}{vector containing two indices of the bands
  to use}
}
\description{
  \code{bandsdata.getbanddistance} gives the distance
  between to bands at one kpoint.
}

