\name{bandsdata.getenergy}
\alias{bandsdata.getenergy}
\title{Gives the energy of a band at one kpoint}
\usage{
  bandsdata.getenergy(bandsdata, kpoint, band)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{kpoint}{index of kpoint to use}

  \item{band}{index of band to use}
}
\description{
  \code{bandsdata.getenergy} gives the energy of a band at
  one kpoint.
}

