\name{bandsdata.getintervallaroundsympoint}
\alias{bandsdata.getintervallaroundsympoint}
\title{Gives a intervall around a high symmetry point}
\usage{
  bandsdata.getintervallaroundsympoint(bandsdata,
    sympointnumber, intervall = 0.2)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{sympointnumber}{number of high symmetry point to
  use}

  \item{intervall}{in per cent}
}
\description{
  \code{bandsdata.getintervallaroundsympoint} gives a
  intervall around a high symmetry point. uses flat (2d)
  kpoints.
}

