\name{bandsfit.dirac.getv}
\alias{bandsfit.dirac.getv}
\title{Calculation of Fermi velocity}
\usage{
  bandsfit.dirac.getv(bandsfit)
}
\arguments{
  \item{bandsfit}{object of class bandsfit, generated by
  dirac fit}
}
\description{
  \code{bandsfit.dirac.getv} calculation of Fermi velocity.
  Will be in 10e6 m/s
}

