\name{bandsfit.quadratic.getm}
\alias{bandsfit.quadratic.getm}
\title{Calculation of effectiv mass}
\usage{
  bandsfit.quadratic.getm(bandsfit)
}
\arguments{
  \item{bandsfit}{object of class bandsfit, generated by
  quadratic fit}
}
\description{
  \code{bandsfit.quadratic.getm} calculation of effectiv
  mass. Will be in units of electron mass
}

