\name{basis.getreciprocal}
\alias{basis.getreciprocal}
\title{Gives reciprocal basis}
\usage{
  basis.getreciprocal(basis, a = 1)
}
\arguments{
  \item{basis}{3x3 matrix}

  \item{a}{(optional) lattice constant which will be
  multiplied with basis}
}
\description{
  \code{poscar.getreciprocalbasis} calculates reciprocal
  basis.
}

