\name{basis.rotate2d}
\alias{basis.rotate2d}
\title{Rotates basis (2d)}
\usage{
  basis.rotate2d(basis, angle = NULL)
}
\arguments{
  \item{basis}{3x3 matrix}

  \item{angle}{in radians}
}
\description{
  \code{basis.rotate2d} rotates basis (2d). Will rotate in
  x-y plane. If no angle is provided. First vector of basis
  is rotated to face in x-direction. Will give 3x3 matrix.
}

