\name{brillouinzone.gethexagonalvector}
\alias{brillouinzone.gethexagonalvector}
\title{Gets hexagonal brilloinzone vector}
\usage{
  brillouinzone.gethexagonalvector(poscar, rotate = 0,
    extend = 1, strain = 0)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{rotate}{rotates brillouinzone (in degrees)}

  \item{extend}{creates supercell of brillouinzone}

  \item{strain}{applied to brillouinzone}
}
\description{
  \code{brillouinzone.gethexagonalvector} gets hexagonal
  brilloinzone vector.
}

