\name{brillouinzone.projectkpoints}
\alias{brillouinzone.projectkpoints}
\title{Gives brillouinzone projected kpoints.}
\usage{
  brillouinzone.projectkpoints(brillouinzone, kpoints)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{kpoints}{object with arbitrary kpoints}
}
\description{
  \code{brillouinzone.projectkpoints} projects arbitrary
  kpoints back in the first brillouinzone.
}

