\name{brillouinzone.selectkpoints}
\alias{brillouinzone.selectkpoints}
\title{Select kpoints in first brillouinzone.}
\usage{
  brillouinzone.selectkpoints(brillouinzone, kpoints,
    maxdistance = 0)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{kpoints}{object with arbitrary kpoints}

  \item{maxdistance}{(optional) allowed distance to first
  brillouinzone (shape-conserving)}
}
\description{
  \code{brillouinzone.selectkpoints} selects kpoints in
  first brillouinzone.
}

