\name{calcrange}
\alias{calcrange}
\title{calculates nested range in list}
\usage{
  calcrange(listofobjects, nameofnestedelement = "range")
}
\arguments{
  \item{listofobjects}{a list of object all containing the
  nameofnestedelement}

  \item{nameofnestedelement}{name of the nested element
  which is in each item of listofobjects}
}
\description{
  \code{calcrange} calculates nested range in list
}

