\name{calculation.getbulkbands}
\alias{calculation.getbulkbands}
\title{Will give a object of class bulkbands}
\usage{
  calculation.getbulkbands(calculation, ...)
}
\arguments{
  \item{calculation}{object of class calculation}
}
\description{
  \code{calculation.getbulkbands} will give a object of
  class bulkbands based on calculation object.
}

