\name{calculation.getea}
\alias{calculation.getea}
\title{Will give a vector containing lattice constants and energies}
\usage{
  calculation.getea(calculation)
}
\arguments{
  \item{calculation}{object of type calculation}
}
\description{
  \code{calculation.getea} Will give a vector containing
  lattice constants and energies based on calculation
  object. Vector will have class ea.
}

