\name{calculation.getindicesfrompositions}
\alias{calculation.getindicesfrompositions}
\title{searches for the indices of the value in the CHGCAR}
\usage{
  calculation.getindicesfrompositions(calculation,
    positions)
}
\arguments{
  \item{calculation}{object of class calculation}

  \item{positions}{object of class positions}
}
\description{
  \code{calculation.getindicesfrompositions} searches for
  the index of the value in the CHGCAR by a given position
}

