\name{dosdata.addsmearing}
\alias{dosdata.addsmearing}
\title{Add smearing to a dosdata object}
\usage{
  dosdata.addsmearing(dosdata, sigma = 0.2)
}
\arguments{
  \item{dosdata}{object of type dosdata}

  \item{sigma}{standard deviation}
}
\description{
  \code{dosdata.addsmearing} Add gaussian smearing to a
  dosdata object.
}

