\name{dosvector.calcsmearing}
\alias{dosvector.calcsmearing}
\title{Add smearing to a dosvector}
\usage{
  dosvector.calcsmearing(energy, dos, sigma = 0.1)
}
\arguments{
  \item{energy}{energy vector}

  \item{dos}{dos vector}

  \item{sigma}{standard deviation}
}
\description{
  \code{dosvector.calcsmearing} add gaussian smearing to a
  dosvector.
}

