\name{file.gethighestversion}
\alias{file.gethighestversion}
\title{Will give the highest version of provided filename}
\usage{
  file.gethighestversion(dir, filename)
}
\arguments{
  \item{dir}{folder to search}

  \item{filename}{filename}
}
\description{
  \code{file.gethighestversion} will give the highest
  version of provided filename. Searches for example for
  POSCAR POSCAR1 POSCAR2 POSCAR3 etc. and will return the
  highest found filename
}

