\name{getplotinplotfig}
\alias{getplotinplotfig}
\title{gets the fig variables of a plotinplot}
\usage{
  getplotinplotfig(x, y, coords = "ndc")
}
\arguments{
  \item{x}{x position}

  \item{y}{y position}

  \item{coords}{coordinate type of positions}
}
\description{
  \code{getplotinplotfig} gets the fig variables of a
  plotinplot
}

