\name{getrasteredparallelogram}
\alias{getrasteredparallelogram}
\title{prints a two python vectors for x and y direction with equidistant points within and on the left and bottom border of a parallelogram}
\usage{
  getrasteredparallelogram(direction1, direction2,
    discretisation1 = 5, discretisation2 = 5,
    significant = 5)
}
\arguments{
  \item{direction1}{first 2d direction}

  \item{direction2}{second 2d direction}

  \item{discretisation1}{number of steps in first
  direction}

  \item{discretisation2}{number of steps in second
  direction}

  \item{significant}{prettier printing by use of signif}
}
\description{
  \code{getrasteredparallelogram} calculates equidistant
  points within a parallelogram
}

