\name{ldosvector.calcsmearing}
\alias{ldosvector.calcsmearing}
\title{Adds a gaussian smearing}
\usage{
  ldosvector.calcsmearing(energy, dos, sigma = 0.1)
}
\arguments{
  \item{energy}{vector containing the energy}

  \item{dos}{vector containing the dos}

  \item{sigma}{smearing parameter}
}
\description{
  \code{ldosvector.calcsmearing} Adds a gaussian smearing
  to ldos vector
}

