\name{load.calculations}
\alias{load.calculations}
\title{Loads a calculations object in RData format}
\usage{
  load.calculations(name, update = F)
}
\arguments{
  \item{name}{calculations object to load}

  \item{update}{if object is already loaded determines, if
  object will be loaded again}
}
\description{
  \code{load.calculations} Loads a calculations object in
  RData format. Will search current working directory for
  \code{name}.RData
}

