\name{makeTransparent}
\alias{makeTransparent}
\title{makes a vector of colors transparent
snippet from stackoverflow.com}
\usage{
  makeTransparent(someColor, alpha = 100)
}
\arguments{
  \item{someColor}{colors}

  \item{alpha}{transparency}
}
\description{
  \code{makeTransparent} makes a vector of colors
  transparent
}

