\name{matrix.reflection2d.degree}
\alias{matrix.reflection2d.degree}
\title{Gets matrix for reflection across a line}
\usage{
  matrix.reflection2d.degree(slope)
}
\arguments{
  \item{slope}{given in degrees}
}
\description{
  \code{matrix.reflection2d} gets matrix for reflection
  across a line which goes through zero.
}

