\name{matrix.reflection2d}
\alias{matrix.reflection2d}
\title{Gets matrix for reflection across a line}
\usage{
  matrix.reflection2d(slope)
}
\arguments{
  \item{slope}{given in radians}
}
\description{
  \code{matrix.reflection2d} gets matrix for reflection
  across a line which goes through zero.
}

