\name{parse.poscar}
\alias{parse.poscar}
\title{Parses a dataframe into object of class poscar}
\usage{
  parse.poscar(stringdataframe)
}
\arguments{
  \item{stringdataframe}{Dataframe of strings}
}
\description{
  \code{parse.poscar} parses a dataframe into object of
  class poscar. The dataframe only consists of one column
  whereas each row contains a row of the poscar.
}

