\name{periodictable.getelementnames}
\alias{periodictable.getelementnames}
\title{gives you the name of elements}
\usage{
  periodictable.getelementnames(element)
}
\arguments{
  \item{element}{vector of (element) numbers}
}
\description{
  \code{periodictable.getelementnames} gives you the name
  of elements.
}

