\name{periodictable.getelementpositions}
\alias{periodictable.getelementpositions}
\title{gives you the positions of specific elements in a periodic table}
\usage{
  periodictable.getelementpositions(element)
}
\arguments{
  \item{element}{vector of (element) numbers or names}
}
\description{
  \code{periodictable.getelementpositions} gives you the
  position of specific elements in a periodic table.
}

