\name{periodictable.getelementselector}
\alias{periodictable.getelementselector}
\title{gives you a selector for specific elements}
\usage{
  periodictable.getelementselector(element)
}
\arguments{
  \item{element}{vector of (element) numbers or names}
}
\description{
  \code{periodictable.getelementselector} gives you a
  selector for specific elements.
}

