\name{periodictable.getremainingelements}
\alias{periodictable.getremainingelements}
\title{gives you a selector for all but specific elements}
\usage{
  periodictable.getremainingelements(element)
}
\arguments{
  \item{element}{vector of (element) numbers or names}
}
\description{
  \code{periodictable.getremainingelements} gives you a
  selector for all but specific elements.
}

