\name{plot.addlabel}
\alias{plot.addlabel}
\title{adds a label to a plot with the help of the package grid}
\usage{
  plot.addlabel(label, top = 0.5, left = 0)
}
\arguments{
  \item{label}{label}

  \item{top}{distance to top in mm}

  \item{left}{distance to left in mm}
}
\description{
  \code{plot.addlabel} adds a label to a plot
}

