\name{plot.atoms.add}
\alias{plot.atoms.add}
\title{Adds atoms to existing plot}
\usage{
  plot.atoms.add(atoms, basis = NULL, direction = 3,
    col = "white", cex = 3, lwd = 1, lty = 1, ...)
}
\arguments{
  \item{atoms}{dataframe of atoms}

  \item{direction}{of projection}

  \item{basis}{basis if atoms are in direct coordinates}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.atoms.add} adds atoms to existing plot.
}

