\name{plot.atoms.addnumbers}
\alias{plot.atoms.addnumbers}
\title{Adds numbers on atoms to existing plot}
\usage{
  plot.atoms.addnumbers(atoms, basis = NULL, direction = 3,
    atomselector = NULL, ...)
}
\arguments{
  \item{atoms}{dataframe of atoms}

  \item{direction}{of projection}

  \item{basis}{basis if atoms are in direct coordinates}

  \item{atomselector}{vector of atom indices which should
  be used}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.atoms.addnumbers} adds numbers on atoms to
  existing plot.
}

