\name{plot.bandsdata.addfermi}
\alias{plot.bandsdata.addfermi}
\title{Adds line at Fermi level to existing plot}
\usage{
  plot.bandsdata.addfermi(bandsdata, col = "blue", lty = 3,
    ...)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{col}{color of line at Fermi level}

  \item{lty}{line typ of line at Fermi level}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.bandsdata.addsymnnames} adds line at Fermi
  level to existing plot
}

