\name{plot.bandsdata.addnumbers}
\alias{plot.bandsdata.addnumbers}
\title{Adds number of all bands to existing plot}
\usage{
  plot.bandsdata.addnumbers(bandsdata, kpoints = NULL,
    energyoffset = NULL, ...)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{kpoints}{at which numbers will be plotted, is
  recycled}

  \item{energyoffset}{will be added to energy of all bands}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.bandsdata.addsymnnames} adds number of all
  bands to existing plot.
}

