\name{plot.bandsdata.addsymnnames}
\alias{plot.bandsdata.addsymnnames}
\title{Adds high symmetry point labels to existing plot}
\usage{
  plot.bandsdata.addsymnnames(bandsdata, symnames,
    symcolor = "red", symlty = 3, ...)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{symnames}{adds labels at high symmetry points}

  \item{symcolor}{color of high symmetry point lines}

  \item{symlty}{line typ of high symmetry point lines}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.bandsdata.addsymnnames} adds high symmetry
  point labels to existing plot
}

