\name{plot.brillouinzone.add}
\alias{plot.brillouinzone.add}
\title{Adds brilloinzonevector to existing plot}
\usage{
  plot.brillouinzone.add(brillouinzone, col = "blue", ...)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.brillouinzone.addvector} adds
  brilloinzonevector to existing plot based on poscar.
}

