\name{plot.brillouinzone.addvector}
\alias{plot.brillouinzone.addvector}
\title{Adds brilloinzonevector to existing plot}
\usage{
  plot.brillouinzone.addvector(brilloinzonevector,
    col = "blue", ...)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.brillouinzone.addvector} adds
  brilloinzonevector to existing plot based on poscar.
}

