\name{plot.brillouinzone}
\alias{plot.brillouinzone}
\title{Plots brillouinzone}
\usage{
  plot.brillouinzone(brillouinzone, xlab = NA, ylab = NA,
    xaxt = "n", yaxt = "n", col = "blue", typ = "l", ...)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.brillouinzone} plots brillouinzone.
}

