\name{plot.brillouinzones.add}
\alias{plot.brillouinzones.add}
\title{Adds brilloinzonevector to existing plot}
\usage{
  plot.brillouinzones.add(brillouinzones, col = "blue",
    ...)
}
\arguments{
  \item{brillouinzones}{object of class brillouinzones}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.brillouinzonse.addvector} adds
  brilloinzonevector to existing plot based on poscar.
}

