\name{plot.brillouinzones}
\alias{plot.brillouinzones}
\title{Plots brillouinzones}
\usage{
  plot.brillouinzones(brillouinzones, xlab = NA, ylab = NA,
    xaxt = "n", yaxt = "n", col = "blue", typ = "l", ...)
}
\arguments{
  \item{brillouinzones}{object of class brillouinzones}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.brillouinzones} plots brillouinzones.
}

