\name{plot.bulkbands.add}
\alias{plot.bulkbands.add}
\title{Adds bulkband polygons to existing plot}
\usage{
  plot.bulkbands.add(bulkbands, col = "grey", ...)
}
\arguments{
  \item{bulkbands}{object of class bulkbands}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.bulkbands.add} adds bulkband polygons to
  existing plot.
}

