\name{plot.calculation.ea.addfit}
\alias{plot.calculation.ea.addfit}
\title{Adds a fit to an existing e over a curve}
\usage{
  plot.calculation.ea.addfit(calculation, afactor = 1,
    energyfactor = 1, energyshift = 0, ...)
}
\arguments{
  \item{calculation}{object of type calculation}

  \item{afactor}{will be multiplied with lattice constant}

  \item{energyfactor}{will be multiplied with energy}

  \item{energyshift}{will be substracted from energy}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.calculation.ea.addfit} adds a fit to an
  existing e over a curve.
}

