\name{plot.dosdata.addfermi}
\alias{plot.dosdata.addfermi}
\title{Will add fermi level to existing plot.}
\usage{
  plot.dosdata.addfermi(dosdata, col = "blue",
    col.text = "blue", lty = 3, ...)
}
\arguments{
  \item{dosdata}{object of type dosdata}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.dosdata.addfermi } will add fermi level to
  existing plot.
}

