\name{plot.EOS.add}
\alias{plot.EOS.add}
\title{Adds a EOS fit to an existing plot}
\usage{
  plot.EOS.add(o, ...)
}
\arguments{
  \item{o}{object of EOS class}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.EOS.add} Adds a EOS fit to an existing plot.
  fits with \code{\link{ea.fitEOS}}.
}

