\name{plot.periodictable.addunderlay}
\alias{plot.periodictable.addunderlay}
\title{Draws rectangular underlayings in existing periodic table}
\usage{
  plot.periodictable.addunderlay(underlay, col = "red")
}
\arguments{
  \item{underlay}{vector of (element) numbers or names}

  \item{col}{color of rectangle underlaying elements}
}
\description{
  \code{plot.periodictable.addunderlay} Draws rectangles
  underlaying selected elements in existing periodic table.
}
\seealso{
  \code{\link{plot.periodictable.addelements}} and
  \code{\link{periodictable.getremainingelements}}
}

