\name{plot.poscar.addlayerdistance}
\alias{plot.poscar.addlayerdistance}
\title{Adds distance between layers to existing plot}
\usage{
  plot.poscar.addlayerdistance(poscar, layer, layers,
    color = "black", direction = 1, ...)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{direction}{of projection}

  \item{layer}{vector of indices of layers to plot}

  \item{layers}{total layer count in poscar}

  \item{color}{color of lines}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.poscar.addlayers} adds distance between layers
  of atoms to existing plot.
}

