\name{plot.poscar.addunitcell}
\alias{plot.poscar.addunitcell}
\title{Adds unitcell to existing plot}
\usage{
  plot.poscar.addunitcell(poscar, direction = 3, lty = 2,
    ...)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{direction}{of projection}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.poscar.addunitcell} adds unitcell to existing
  plot.
}

