\name{plot.positions.addlegend}
\alias{plot.positions.addlegend}
\title{adds a positions legend to an existing plot.}
\usage{
  plot.positions.addlegend(positions, lty = 1,
    col = rainbow(length(positions)),
    position = "topright", ...)
}
\arguments{
  \item{positions}{object of class positions}

  \item{...}{further ploting parameters}
}
\description{
  \code{plot.positions.addlegend} adds a positions legend
  to an existing plot
}

