\name{plot.stm.addunitcell}
\alias{plot.stm.addunitcell}
\title{Adds unitcell}
\usage{
  plot.stm.addunitcell(stm, atomnumber = NULL,
    col = "black", ...)
}
\arguments{
  \item{stm}{object of class stm}

  \item{atomnumber}{(optional) number of the atom on which
  the bottom left corner of the unitcell will be
  positioned}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.stm.addunitcell} Adds the unitcell to a stm
  image indicated by a dashed line
}

