\name{plot.stm}
\alias{plot.stm}
\title{Creates a constant-current stm image.
Based on lattice package.}
\usage{
  plot.stm(stm, super = 1, xlim = NULL, ylim = NULL, ...)
}
\arguments{
  \item{stm}{object of class stm}

  \item{super}{(optional) plots a supercell (symmetric in x
  and y)}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.stm} Plots a constant-current stm image
}

