\name{plotinplot}
\alias{plotinplot}
\title{creates a new plot within a plot
uses base package}
\usage{
  plotinplot(x, y, mar = c(0, 0, 0, 0))
}
\arguments{
  \item{x}{x position}

  \item{y}{y position}

  \item{mar}{margin for new plot}
}
\description{
  \code{plotinplot} creates a new plot within a plot
}

