\name{poscar.changebasis}
\alias{poscar.changebasis}
\title{Changes the basis of a poscar}
\usage{
  poscar.changebasis(poscar, newbasis)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{newbasis}{3x3 matrix containing new basis vectors
  as rows}
}
\description{
  \code{poscar.changebasis} changes the basis of a poscar.
}

