\name{poscar.getbasisconvertedatoms}
\alias{poscar.getbasisconvertedatoms}
\title{Calculates basis converted atoms}
\usage{
  poscar.getbasisconvertedatoms(poscar, forth = F)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{forth}{from cartesian to direct}
}
\description{
  \code{poscar.getbasisconvertedatoms} calculates basis
  converted atoms.
}

