\name{poscar.getbrillouinzone.hexagonal}
\alias{poscar.getbrillouinzone.hexagonal}
\title{Gets hexagonal brilloinzone vector}
\usage{
  poscar.getbrillouinzone.hexagonal(poscar, rotate = 0,
    extend = 1, strain = 0)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{rotate}{rotates brillouinzone (in degrees)}

  \item{extend}{creates supercell of brillouinzone}

  \item{strain}{applied to brillouinzone}
}
\description{
  \code{poscar.getbrillouinzone.hexagonal} gets hexagonal
  brilloinzone vector.
}

