\name{poscar.getbrillouinzone.rectangular}
\alias{poscar.getbrillouinzone.rectangular}
\title{Gets rectangular brilloinzone vector}
\usage{
  poscar.getbrillouinzone.rectangular(poscar, rotate = 0,
    extend = 1, strain = 0)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{rotate}{rotates brillouinzone (in degrees)}

  \item{extend}{creates supercell of brillouinzone}

  \item{strain}{applied to brillouinzone}
}
\description{
  \code{poscar.getbrillouinzone.rectangular} gets
  rectangular brilloinzone vector.
}

