\name{poscar.getshapedposcar}
\alias{poscar.getshapedposcar}
\title{gives you a poscar with atoms in area of given size and shape}
\usage{
  poscar.getshapedposcar(poscar, x, y,
    shape = c("rectangular"))
}
\arguments{
  \item{poscar}{the input poscar.}

  \item{x}{max x (range will be c(0,x))}
}
\description{
  \code{poscar.getshapedposcar} test
}

