\name{poscar.getvacuum}
\alias{poscar.getvacuum}
\title{Gives the vacuum}
\usage{
  poscar.getvacuum(poscar)
}
\arguments{
  \item{poscar}{object of class poscar}
}
\description{
  \code{poscar.getvacuum} calculates the vacuum. Vacuum is
  defined by the distance between atoms across unitcell
  borders.
}

