\name{poscar.mirrorlayers}
\alias{poscar.mirrorlayers}
\title{gives you a poscar with all atoms in mirrorlayer
mirrored by the diagonal going trough a atom of the baselayer}
\usage{
  poscar.mirrorlayers(poscar, layers, baselayer,
    mirrorlayers)
}
\arguments{
  \item{poscar}{the input poscar.}

  \item{layers}{layercount of poscar}

  \item{baselayer}{layer in which the mirror-diagonal will
  be layed}

  \item{mirrorlayers}{layers which will be mirrored by the
  diagonal}
}
\description{
  \code{poscar.mirrorlayers} test
}

