\name{poscar.rotate2d}
\alias{poscar.rotate2d}
\title{Rotates basis (2d)}
\usage{
  poscar.rotate2d(poscar, angle = NULL)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{angle}{in radians}
}
\description{
  \code{poscar.rotate2d} rotates basis (2d). Will rotate in
  x-y plane If no angle is provided. First vector of basis
  is rotated to face in x-direction. Will give object of
  class poscar
}

