\name{poscar.sortatoms}
\alias{poscar.sortatoms}
\title{Sorts atoms in object of class poscar}
\usage{
  poscar.sortatoms(poscar, sortindices = c(7, 3, 1, 2))
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{sortindices}{vector of indices after which to sort.
  \code{c(1,3)} will sort by rx then by rz}
}
\description{
  \code{poscar.sortatoms} sorts atoms in object of class
  poscar
}

