\name{poscar.translatelayers}
\alias{poscar.translatelayers}
\title{Translates layer by given 2d translation}
\usage{
  poscar.translatelayers(poscar, layer, layers,
    directtranslation)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{layer}{indices of layers which will be moved}

  \item{layers}{total layer count}

  \item{directtranslation}{2d translational vector}
}
\description{
  \code{poscar.translatelayers} translates layer by given
  2d translation. Will translate in x-y plain
}

