\name{predict.bandsfit}
\alias{predict.bandsfit}
\title{Custom predict for object of class bandsfit}
\usage{
  predict.bandsfit(o, x)
}
\arguments{
  \item{o}{object}

  \item{x}{values to use for prediction}
}
\description{
  \code{predict.bandsfit} custom predict for object of
  class bandsfit.
}

