\name{predict.EOS}
\alias{predict.EOS}
\title{Predicts values based on a equation of state fit}
\usage{
  predict.EOS(o, x)
}
\arguments{
  \item{o}{object of EOS type}

  \item{x}{values for which a y should be predicted}
}
\description{
  \code{predict.EOS} Predicts values based on a equation of
  state fit.
}

