\name{print.calculation}
\alias{print.calculation}
\title{Custom print for objects of class calculation}
\usage{
  print.calculation(calculation, ...)
}
\arguments{
  \item{calculation}{objects of class calculation}
}
\description{
  \code{print.calculation} custom print for objects of
  class calculation.
}

