\name{print.calculations}
\alias{print.calculations}
\title{Custom print for objects of class calculations}
\usage{
  print.calculations(calculations, ...)
}
\arguments{
  \item{calculations}{objects of class calculations}
}
\description{
  \code{print.calculations} custom print for objects of
  class calculations.
}

