\name{reciprocalbasis.getbrillouinzone}
\alias{reciprocalbasis.getbrillouinzone}
\title{Gets brillouinzone vector}
\usage{
  reciprocalbasis.getbrillouinzone(recbasis)
}
\arguments{
  \item{recbasis}{basis in reciprocal space}
}
\description{
  \code{reciprocalbasis.getbrillouinzone} gets
  brillouinzone vector.
}

