\name{reciprocalbasis.getbrillouinzonetype}
\alias{reciprocalbasis.getbrillouinzonetype}
\title{Gets type of brillouinzone}
\usage{
  reciprocalbasis.getbrillouinzonetype(recbasis)
}
\arguments{
  \item{recbasis}{basis in reciprocal space}
}
\description{
  \code{reciprocalbasis.getbrillouinzone} gets type of
  brillouinzone.
}

