\name{vectors.arecollinear}
\alias{vectors.arecollinear}
\title{Checks two vectors for collinearity}
\usage{
  vectors.arecollinear(vector1, vector2, prec = 1e-06)
}
\arguments{
  \item{vector1}{first vector}

  \item{vector2}{second vector}

  \item{prec}{(optional) precission}
}
\description{
  \code{vectors.arecollinear} checks two vectors for
  collinearity.
}

