\name{vectors.calcangle.degree}
\alias{vectors.calcangle.degree}
\title{Gets the angle between two vectors}
\usage{
  vectors.calcangle.degree(vector1, vector2, period = NA)
}
\arguments{
  \item{vector1}{first vector}

  \item{vector2}{second vector}

  \item{period}{(optional) if provided folds the angle back
  to its period. Period in degrees.}
}
\description{
  \code{vectors.calcangle.degree} calcultes the angle
  between two vectors. Gives angle in degrees.
}

