\name{vectors.calcangle}
\alias{vectors.calcangle}
\title{Gets the angle between two vectors}
\usage{
  vectors.calcangle(vector1, vector2, period = NA)
}
\arguments{
  \item{vector1}{first vector}

  \item{vector2}{second vector}

  \item{period}{(optional) if provided folds the angle back
  to its period. Period in radians.}
}
\description{
  \code{vectors.calcangle} calcultes the angle between two
  vectors. Gives angle in radians.
}

