\name{vectors.crossproduct}
\alias{vectors.crossproduct}
\title{Calculated the crossproduct of two given 3d vectors}
\usage{
  vectors.crossproduct(vector1, vector2)
}
\arguments{
  \item{vector1}{first vector}

  \item{vector2}{second vector}
}
\description{
  \code{crossprod.vec} calculated the crossproduct
}

