\name{write.poscar}
\alias{write.poscar}
\title{Writes object of class poscar into file of POSCAR format}
\usage{
  write.poscar(poscar, file)
}
\arguments{
  \item{poscar}{object of class poscar}

  \item{file}{name of file}
}
\description{
  \code{write.poscar} writes object of class poscar into
  file of POSCAR format.
}

