\name{brillouinzone.extendkpoints}
\alias{brillouinzone.extendkpoints}
\title{Extends kpoints two all second brillouinzones.}
\usage{
  brillouinzone.extendkpoints(brillouinzone, kpoints)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{kpoints}{object with arbitrary kpoints}
}
\description{
  \code{brillouinzone.extendkpoints} translates kpoints two
  all second brillouinzones.
}

