\name{brillouinzone.selectkpoints}
\alias{brillouinzone.selectkpoints}
\title{Select kpoints in first brillouinzone.}
\usage{
  brillouinzone.selectkpoints(brillouinzone, kpoints)
}
\arguments{
  \item{brillouinzone}{object of class brillouinzone}

  \item{kpoints}{object with arbitrary kpoints}
}
\description{
  \code{brillouinzone.selectkpoints} selects kpoints in
  first brillouinzone.
}

