\name{plot.bandsdata.addsymnnames}
\alias{plot.bandsdata.addsymnnames}
\title{Adds high symmetry point labels to existing plot}
\usage{
  plot.bandsdata.addsymnnames(bandsdata, labels,
    col.labels = "black", lty = 3, col.lines = "red", ...)
}
\arguments{
  \item{bandsdata}{object of class bandsdata}

  \item{labels}{adds labels at high symmetry points}

  \item{col.labels}{color labels}

  \item{lty}{line typ of high symmetry point lines}

  \item{col.lines}{color of high symmetry point lines}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.bandsdata.addsymnnames} adds high symmetry
  point labels to existing plot
}

