\name{plot.stm}
\alias{plot.stm}
\title{Plots a constant-current stm image}
\usage{
  plot.stm(stm, super = 1, xlim = NULL, ylim = NULL,
    more = F, ...)
}
\arguments{
  \item{stm}{object of class stm}

  \item{super}{(optional) plots a supercell (symmetric in x
  and y)}

  \item{...}{further plotting parameters}
}
\description{
  \code{plot.stm} Creates a constant-current stm image.
  Based on lattice package.
}

