\name{reciprocalbasis.getbrillouinzone}
\alias{reciprocalbasis.getbrillouinzone}
\title{Gets brillouinzone vector}
\usage{
  reciprocalbasis.getbrillouinzone(recbasis,
    zone = c("bz", "basis"))
}
\arguments{
  \item{recbasis}{basis in reciprocal space}

  \item{zone}{specific the wished zone. \code{bz} will
  calculate the first brillouinzone, \code{basis} will give
  the simple zone conducted by the reciprocal basis
  vectors}
}
\description{
  \code{reciprocalbasis.getbrillouinzone} gets
  brillouinzone vector.
}

